Finished
As we could see that 3 sum is far more challanging because you have to
set the target first. Then sort the whole vector to make sure you would
not have to worry about the duplicate. Then you have to set two pointer
front and back to narrow the limitation. Remeber the duplication
limitation is the key.